requirements.txt
anyio==4.0.0
asgiref==3.7.2
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
dj-database-url==2.1.0
Django==4.2.6
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
django-filter==23.3
django-google-sso==3.3.0
django-widget-tweaks==1.5.0
exceptiongroup==1.1.3
google-auth==2.23.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
idna==3.4
loguru==0.7.2
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6
whitenoise==6.5.0
